backend/requirements.txt
Django
gunicorn
djangorestframework
coreapi
django-filter
django-crispy-forms
djoser
django-cors-headers
huggingface_hub